NEW add global Recipients tab. remove all 'Delete' buttons where ever suitable, add couple of "Archive" buttons and use "GridFieldArchiveAction" where ever suitable.

+++ tasks/DeleteNewsletterArchiveObjectTask.php
@@ -51,7 +51,7 @@ class DeleteNewsletterArchivedObjectTask extends DailyTask{
 		$count = $objects->count();
 		if($count){
 			foreach($objects as $object){
-				$object->delete();
+				if($object->canDelete()) $object->delete();
 			}
 		}
 		return $count;
